CDN/requirements.txt
firebase-admin
flask
pyjwt
bcrypt